Refactor custom rules architecture for clarity
