+++ backend/src/middlewares/fishtank.js
@@ -1,8 +1,8 @@
 const { checkSchema } = require('express-validator/check');
-const User = require('./validation/customRules/checkers/user.js');
-const Shoal = require('./validation/customRules/checkers/shoal.js');
+const User = require('./validation/customRules/common/user.js');
+const Shoal = require('./validation/customRules/common/shoal.js');
 const ValidationSchema = require('./validation/schemas/fishtank.js');
-const checkCustomRules = require('./validation/customRules/checkers/fishtank.js');
+const CustomRules = require('./validation/customRules/fishtank.js');
 const bail = require('./validation/errorHandler.js');
 
 module.exports = {
@@ -19,7 +19,7 @@ module.exports = {
     bail(403),
     checkSchema(ValidationSchema.edit),
     bail(422),
-    checkCustomRules.edit,
+    CustomRules.edit,
     bail(422),
     User.isOwner,
     bail(403),
