--- backend/src/middlewares/validation/customRules/checkers/fishtank.js
@@ -1,9 +1,0 @@
-const { param } = require('express-validator/check');
-const Rules = require('../schemas/fishtank.js');
-
-module.exports = {
-  edit: param('id').custom(Rules.exists.options)
-    .withMessage(Rules.exists.errorMessage)
-    .custom(Rules.isOngoing.options)
-    .withMessage(Rules.isOngoing.errorMessage),
-};
