--- backend/src/middlewares/validation/customRules/checkers/shoal.js
@@ -1,6 +1,0 @@
-const { body } = require('express-validator/check');
-const { isValidShoal } = require('../../../../__mock_teamy__');
-
-module.exports = {
-  isValid: body('shoalId').custom(isValidShoal),
-};
